requirements.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
Django==3.1.3
django-bulma==0.8.1
django-common-helpers==0.9.2
django-cron==0.5.1
django-modeltranslation==0.16
django-paypal==1.0.0
django-tinymce==3.1.0
idna==2.10
mysqlclient==2.0.1
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
wincertstore==0.2
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.2.0
